+++ train/src/matrix.cpp
@@ -12,7 +12,7 @@ namespace entity2vec{
         data_ = nullptr;
     }
 
-    matrix::matrix(uint64_t m, uint64_t n) {
+    matrix::matrix(uint32_t m, uint32_t n) {
         m_ = m;
         n_ = n;
         data_ = new real[m * n];
@@ -21,27 +21,27 @@ namespace entity2vec{
     void matrix::uniform(real a) {
         std::minstd_rand rng(1);
         std::uniform_real_distribution<> uniform(-a, a);
-        for (uint64_t i = 0; i < (m_ * n_); i++) {
+        for (uint32_t i = 0; i < (m_ * n_); i++) {
             data_[i] = uniform(rng);
         }
     }
 
     void matrix::zero() {
-        for (uint64_t i = 0; i < (m_ * n_); i++) {
+        for (uint32_t i = 0; i < (m_ * n_); i++) {
             data_[i] = 0.0;
         }
     }
 
-    real matrix::dotRow(const vector &vec, uint64_t i) {
+    real matrix::dotRow(const vector &vec, uint32_t i) {
         real d = 0.0;
-        for (uint64_t j = 0; j < n_; j++) {
+        for (uint32_t j = 0; j < n_; j++) {
             d += data_[i * n_ + j] * vec.data_[j];
         }
         return d;
     }
 
-    void matrix::addRow(const vector &vec, uint64_t i, real a) {
-        for (uint64_t j = 0; j < n_; j++) {
+    void matrix::addRow(const vector &vec, uint32_t i, real a) {
+        for (uint32_t j = 0; j < n_; j++) {
             data_[i * n_ + j] += a * vec.data_[j];
         }
     }
