+++ train/src/util.cpp
@@ -54,7 +54,7 @@ namespace util{
         initSigmoid();
     }
 
-    void seek(std::ifstream &ifs, uint64_t pos) {
+    void seek(std::ifstream &ifs, uint32_t pos) {
         ifs.clear();
         ifs.seekg(std::streampos(pos));
     }
